- Buildfix for libnx - Remove now unused p_rarch pointers in functions
